Fixed test import issue
